+++ pyobs/modules/focus/projection.py
@@ -5,7 +5,7 @@ import numpy as np
 from scipy import optimize, ndimage
 
 from pyobs.comm import RemoteException
-from pyobs.interfaces import IFocuser, ICamera, IAutoFocus, IFilters, ICameraExposureTime, IImageType
+from pyobs.interfaces import IFocuser, ICamera, IAutoFocus, IFilters, IExposureTime, IImageType
 from pyobs.events import FocusFoundEvent
 from pyobs.modules import Module
 from pyobs.modules import timeout
@@ -134,7 +134,7 @@ class AutoFocusProjection(Module, IAutoFocus):
             if self._abort.is_set():
                 raise InterruptedError()
             try:
-                if isinstance(camera, ICameraExposureTime):
+                if isinstance(camera, IExposureTime):
                     camera.set_exposure_time(exposure_time)
                 if isinstance(camera, IImageType):
                     camera.set_image_type(ImageType.FOCUS)